Stats now use arrays instead of single values.
After pondering it a while, I changed the structure of the skill stats
like barter to use arrays instead. While it means we'll probably need to
write dedicated accessor and mutator methods for each of those
properties so that other objects look at the sum instead of the array
directly, it also means that things like level-ups and skill-point
assignment don't need to spaz out recalculating the stat directly. Just
change the base (index 0) with level ups and ranks (index 2) when
skillpoints are assigned after level up. And tags would be assigned
during character creation or when taking certain perks to a value of 15.

Also, added Level Up button.