Name Generator Works
Implemented a basic form of the name generator that doesn't take into
account any of the features, strengths, or tendencies of the unit.
Implemented there different name dictionaries to create random names
from. Added buttons to create and display random names on the main page.
Changed the style of the main page so that there is a scroll window.
Reused old code from a previous project so that the new scrollable div
will also jump to the bottom whenever it is updated. Put those options
to manage the div or add content to it to the Utilities object. as
static methods.