Changed Namegen Arguments Order
Namegen now takes the number of names to generate as the first argument,
and the list of names to use is created from all following arguments.
This will make it easier to have a toggle switch for which dictionaries
to use, and then just hit a button to use toggled dictionaries.