Dictionary Toggles And New Button Layout
Implemented a system such that dictionaries can be toggled on and off by
clicking their buttons. Changed the layout so that the Newgen button and
the toggle buttons are not on the same line. Added handles for all
buttons in the handleUI javascript file. Units now write levelup
messages to the main output window. Units also now get their randomly
generated names. Test was updated to also generate a set of FoE Vanilla
names.